# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "STAC api written in go with fiber and mongodb",
        "title": "go-stac-api",
        "contact": {
            "name": "Jonathan Healy",
            "email": "jonathan.d.healy@gmail.com"
        },
        "version": "0.0"
    },
    "host": "localhost:6001",
    "basePath": "/",
    "paths": {
        "/collections": {
            "get": {
                "description": "Get all Collections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Get all Collections",
                "operationId": "get-all-collections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Collection"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a collection with a unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Create a STAC collection",
                "operationId": "post-collection",
                "parameters": [
                    {
                        "description": "STAC Collection json",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/collections/{collectionId}": {
            "get": {
                "description": "Get a collection by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Get a Collection",
                "operationId": "get-collection-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit a collection by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Edit a Collection",
                "operationId": "edit-collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "STAC Collection json",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a collection by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Delete a Collection",
                "operationId": "delete-collection-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/collections/{collectionId}/items": {
            "get": {
                "description": "Get all Items with a Collection ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ItemCollection"
                ],
                "summary": "Get all Items from a Collection",
                "operationId": "get-item-collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ItemCollection"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an item with an ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Create a STAC item",
                "operationId": "post-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "STAC Item json",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/collections/{collectionId}/items/{itemId}": {
            "get": {
                "description": "Get an item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get an item",
                "operationId": "get-item-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit a stac item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Edit an Item",
                "operationId": "edit-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "STAC Collection json",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an Item by ID is a specified collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Delete an Item",
                "operationId": "delete-item-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/search": {
            "post": {
                "description": "Search for STAC items via the Search endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "POST Search request",
                "operationId": "post-search",
                "parameters": [
                    {
                        "description": "Search body json",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Search"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Collection": {
            "type": "object",
            "properties": {
                "crs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "extent": {},
                "id": {
                    "type": "string"
                },
                "itemType": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "license": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Link"
                    }
                },
                "providers": {
                    "type": "array",
                    "items": {}
                },
                "stac_version": {
                    "type": "string"
                },
                "summary": {},
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Context": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "returned": {
                    "type": "integer"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "assets": {},
                "bbox": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "collection": {
                    "type": "string"
                },
                "geometry": {},
                "id": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {}
                },
                "properties": {},
                "stac_extensions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stac_version": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.ItemCollection": {
            "type": "object",
            "properties": {
                "context": {
                    "$ref": "#/definitions/models.Context"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Link": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                },
                "rel": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Search": {
            "type": "object",
            "properties": {
                "collections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                }
            }
        }
    }
}